pyproject.toml
[project]
name = "tutorials"
version = "0.1.0"
description = "Automated tutorial factory pipeline: generate narrated tutorial videos from a single function call"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "azure-cognitiveservices-speech>=1.30.0",
    "faster-whisper>=1.0.0",
    "ipykernel>=7.0.0",
    "ipywidgets>=8.0.0",
    "ruff>=0.4.0",
    "tqdm>=4.60.0",
    "pytest>=8.0.0",
    "azure-identity>=1.25.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
